Rename serde modules to pod and move under separate pod feature
