+++ tiledb/api/Cargo.toml
@@ -38,6 +38,7 @@ tiledb-utils = { workspace = true }
 [features]
 default = []
 arrow = ["dep:arrow", "dep:serde", "dep:serde_json", "tiledb-common/arrow", "tiledb-common/serde"]
+pod = ["dep:tiledb-pod"]
 proptest-strategies = ["dep:cells", "dep:proptest"]
 serde = ["dep:serde", "dep:serde_json", "dep:tiledb-pod"]
